Update README.md
When using the strict version require of 3.0.0 you won't get the latest version for Symfony 4. I noticed this, because the docs say:
```
By using $resourceCollection->getQuery() you can get access on the query object. use "r" alias for referring to the current entity. in this example the "r" refers to the "ProjectEntity"
```
But the code in tag 3.0.0 did not have the generateQuery() inside the setRepository() function. So normally getQuery returns a queryBuilder, but without the 'generateQuery()' function, getQuery will allways return NULL and therefore it's impossible to add a where condition to a query.
```
/**
     * Sets the Repository And makes query_object.
     *
     * @param entityRepository $repository
     *                                     The Repository
     */
    public function setRepository(EntityRepository $repository): void
    {
        $this->repository = $repository;

        $this->query = $this->generateQuery();
    }
```
Can you fix this in de docs?